fixed Delete repository
git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/trunk@135 72d1a826-0e72-11de-b5bf-434079a3519f

+++ src/java/xc/mst/manager/repository/DefaultProviderService.java
@@ -85,7 +85,9 @@ public class DefaultProviderService implements ProviderService{
     	// Delete schedule for this repository
     	ScheduleService scheduleService = new DefaultScheduleService();
     	HarvestSchedule harvestSchedule = scheduleService.getScheduleForProvider(provider);
-    	scheduleService.deleteSchedule(harvestSchedule);
+    	if (harvestSchedule != null) {
+    		scheduleService.deleteSchedule(harvestSchedule);
+    	}
     	
     	// Delete processing directive for this repository
     	ProcessingDirectiveService processingDirectiveService = new DefaultProcessingDirectiveService();
